Vote Update and Delete only by Creator

+++ src/main/java/com/axiomq/starwars/services/impl/VoteServiceImpl.java
@@ -3,6 +3,7 @@ package com.axiomq.starwars.services.impl;
 import com.axiomq.starwars.entities.Character;
 import com.axiomq.starwars.entities.User;
 import com.axiomq.starwars.entities.Vote;
+import com.axiomq.starwars.exceptions.ObjectNotFoundException;
 import com.axiomq.starwars.repositories.VoteRepository;
 import com.axiomq.starwars.services.CharacterService;
 import com.axiomq.starwars.services.UserService;
@@ -19,7 +20,6 @@ import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.security.Principal;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.UUID;
 
 @RequiredArgsConstructor
@@ -60,15 +60,19 @@ public class VoteServiceImpl implements VoteService {
     @Override
     public Vote getVoteById(Long id) {
         return voteRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException(String.format("Vote not found: %d", id)));
+                .orElseThrow(() -> new ObjectNotFoundException(String.format("Vote not found: %d", id)));
     }
 
     @Override
     public Vote updateVote(Vote vote, MultipartFile icon, Long id, Principal principal) throws IOException{
         Vote existing = getVoteById(id);
 
+        if(!existing.getUser().getEmail().equals(principal.getName()))
+            throw new ObjectNotFoundException("You are not authorized for this action.");
+
         existing.setComment(vote.getComment());
         existing.setValue(vote.getValue());
+
         Path path = saveFile(vote, icon);
         existing.setIcon(path.getFileName().toString());
         existing.setUrl(path.toString());
@@ -79,6 +83,8 @@ public class VoteServiceImpl implements VoteService {
     @Override
     public void deleteVote(Long id, Principal principal) {
         Vote vote = getVoteById(id);
+        if(!vote.getUser().getEmail().equals(principal.getName()))
+            throw new ObjectNotFoundException("You are not authorized for this action.");
 
         Character character = vote.getCharacter();
         voteRepository.deleteById(vote.getId());
